Update all implicit type casts to be explicit

+++ problems/p02/p02.mojo
@@ -34,8 +34,8 @@ def main() raises:
         expected.enqueue_fill(0)
         with a.map_to_host() as a_host, b.map_to_host() as b_host:
             for i in range(SIZE):
-                a_host[i] = i
-                b_host[i] = i
+                a_host[i] = Float32(i)
+                b_host[i] = Float32(i)
                 expected[i] = a_host[i] + b_host[i]
 
         ctx.enqueue_function[add, add](
