+++ problems/p05/p05.mojo
@@ -34,8 +34,8 @@ def main() raises:
         b.enqueue_fill(0)
         with a.map_to_host() as a_host, b.map_to_host() as b_host:
             for i in range(SIZE):
-                a_host[i] = i + 1
-                b_host[i] = i * 10
+                a_host[i] = Float32(i + 1)
+                b_host[i] = Float32(i * 10)
 
             for i in range(SIZE):
                 for j in range(SIZE):
