+++ solutions/p26/p26.mojo
@@ -65,7 +65,7 @@ def butterfly_parallel_max[
         # Start with half the warp size and reduce by half each step
         var offset = WARP_SIZE // 2
         while offset > 0:
-            max_val = max(max_val, shuffle_xor(max_val, offset))
+            max_val = max(max_val, shuffle_xor(max_val, UInt32(offset)))
             offset //= 2
 
         # All threads now have the maximum value across the entire warp
@@ -103,10 +103,10 @@ def butterfly_conditional_max[
         # Butterfly reduction for both maximum and minimum: dynamic for any WARP_SIZE
         var offset = WARP_SIZE // 2
         while offset > 0:
-            var neighbor_val = shuffle_xor(current_val, offset)
+            var neighbor_val = shuffle_xor(current_val, UInt32(offset))
             current_val = max(current_val, neighbor_val)
 
-            var min_neighbor_val = shuffle_xor(min_val, offset)
+            var min_neighbor_val = shuffle_xor(min_val, UInt32(offset))
             min_val = min(min_val, min_neighbor_val)
 
             offset //= 2
@@ -211,7 +211,7 @@ def warp_partition[
         # Butterfly reduction to get total across the warp: dynamic for any WARP_SIZE
         var offset = WARP_SIZE // 2
         while offset > 0:
-            warp_left_total += shuffle_xor(warp_left_total, offset)
+            warp_left_total += shuffle_xor(warp_left_total, UInt32(offset))
             offset //= 2
 
         # Phase 4: Write to output positions
@@ -235,7 +235,7 @@ def test_butterfly_pair_swap() raises:
 
         with input_buf.map_to_host() as input_host:
             for i in range(SIZE):
-                input_host[i] = i
+                input_host[i] = Float32(i)
 
         var input_tensor = LayoutTensor[dtype, layout, ImmutAnyOrigin](
             input_buf
@@ -261,10 +261,10 @@ def test_butterfly_pair_swap() raises:
         for i in range(SIZE):
             if i % 2 == 0:
                 # Even positions get odd values
-                expected_buf[i] = i + 1
+                expected_buf[i] = Float32(i + 1)
             else:
                 # Odd positions get even values
-                expected_buf[i] = i - 1
+                expected_buf[i] = Float32(i - 1)
 
         with output_buf.map_to_host() as output_host:
             print("output:", output_host)
@@ -284,7 +284,7 @@ def test_butterfly_parallel_max() raises:
 
         with input_buf.map_to_host() as input_host:
             for i in range(SIZE):
-                input_host[i] = i * 2
+                input_host[i] = Float32(i * 2)
             # Make sure we have a clear maximum
             input_host[SIZE - 1] = 1000.0
 
@@ -330,9 +330,9 @@ def test_butterfly_conditional_max() raises:
             for i in range(SIZE_2):
                 if i < 9:
                     var values = [3, 1, 7, 2, 9, 4, 8, 5, 6]
-                    input_host[i] = values[i]
+                    input_host[i] = Float32(values[i])
                 else:
-                    input_host[i] = i % 10
+                    input_host[i] = Float32(i % 10)
 
         var input_tensor = LayoutTensor[dtype, layout_2, ImmutAnyOrigin](
             input_buf
@@ -392,7 +392,7 @@ def test_warp_inclusive_prefix_sum() raises:
 
         with input_buf.map_to_host() as input_host:
             for i in range(SIZE):
-                input_host[i] = i + 1
+                input_host[i] = Float32(i + 1)
 
         var input_tensor = LayoutTensor[dtype, layout, ImmutAnyOrigin](
             input_buf
@@ -459,7 +459,7 @@ def test_warp_partition() raises:
                 13,
             ]
             for i in range(SIZE):
-                input_host[i] = test_values[i % len(test_values)]
+                input_host[i] = Float32(test_values[i % len(test_values)])
 
         var input_tensor = LayoutTensor[dtype, layout, ImmutAnyOrigin](
             input_buf
